Handle JwtDescriptor Header & Payload setter/merge.
Payload is a classical get/set property without side effect:
it is easy to reuse the same JwsDescriptor (and its header)
for multiple payloads.

Header is "initializable": it is preconfigured (from ctors), supports
the initialization syntax but cannot be set to another Header
instance (no "merge" side effect).
This is done thanks to the C# 9 init magic.